Added methods on coverting Image to String, and Date to Int16
