+++ Water My Plants/Water My Plants.xcodeproj/xcuserdata/bohdantkachenko.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,7 @@
 		<key>Water My Plants.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>12</integer>
+			<integer>18</integer>
 		</dict>
 	</dict>
 </dict>
