+++ Water My Plants/Water My Plants/View Controllers/AddPlantViewController.swift
@@ -23,10 +23,7 @@ class AddPlantViewController: UIViewController {
     
     private var countDownTime: UIDatePicker?
     var imagePicker = UIImagePickerController()
-    
-    
-    
-    
+
     // MARK: - View Loading
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -36,7 +33,7 @@ class AddPlantViewController: UIViewController {
     
     override func viewWillDisappear(_ animated: Bool) {
         super.viewWillDisappear(animated)
-       
+        
     }
     
     // MARK: - Private
@@ -53,15 +50,15 @@ class AddPlantViewController: UIViewController {
     }
     
     @objc func dateChanged(datePicker: UIDatePicker) {
-          let formatter = DateFormatter()
-          formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
-          
-          h20FrequencyTextField.text = formatter.string(from: datePicker.date)
-      }
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
+        
+        h20FrequencyTextField.text = formatter.string(from: datePicker.date)
+    }
     
     @objc func viewTapped(gestureReconizer: UITapGestureRecognizer) {
-           view.endEditing(true)
-       }
+        view.endEditing(true)
+    }
     
     // MARK: -  Actions
     
@@ -71,42 +68,23 @@ class AddPlantViewController: UIViewController {
         present(imagePicker, animated: true, completion: nil)
     }
     
-    private func convertImageToBase64String () -> String {
-        
-        guard let image = imageView.image else { return "" }
-        return image.jpegData(compressionQuality: 1)?.base64EncodedString() ?? ""
-        
-     }
-    
+
     @IBAction func saveButtonTapped(_ sender: Any) {
         //Saving data to server
         guard let nickname = nicknameTextField.text, !nickname.isEmpty,
-            let date = h20FrequencyTextField.text,
-            let species = speciesTextField.text else { return }
-        
-        let formatterToString = DateFormatter()
-        formatterToString.timeZone = .current
-        formatterToString.dateFormat = "yyyy-MM-dd HH:mm:ss"
-        
+              let date = h20FrequencyTextField.text,
+              let species = speciesTextField.text,
+              let image = imageView.image else { return }
         
         let formatterToDateFromString = DateFormatter()
         formatterToDateFromString.timeZone = .current
         
- 
-       
-       //   Msg_Date_ =  dateFormatterPrint.string(from: datee ?? Date())
-
-        //  print(Msg_Date_)
-        
-        
-        
         let timeDate = formatterToDateFromString.date(from: date)
-        
-//        guard let timeInterval = formatterToDateFromString.date(from: date)?.timeIntervalSince1970 else { return }
-       
+        print(timeDate)
         // convert to Integer
-        let myInt = Int(timeDate?.intVal ?? 1)
-        let plant = Plant(nickName: nickname, h2oFrequency: Int16(myInt), imageUrl: convertImageToBase64String(), species: species)
+        let myInt = Int16(timeDate?.intVal ?? 1)
+        
+        let plant = Plant(nickName: nickname, h2oFrequency: myInt, imageUrl: EncodeDecodeImage.convertImageToBase64String(img: image), species: species)
         
         plantController.sendPlantToServer(plant: plant)
         
@@ -117,16 +95,15 @@ class AddPlantViewController: UIViewController {
         } catch {
             NSLog("Error saving managed object context with error: \(error)")
         }
-
+        
         localNotifHelper.requestAuthorizationStatus { success in
             if success == true {
                 self.localNotifHelper.scheduleDailyReminderNotification(name: nickname, times: Date(), calendar: Calendar.current)
             }
         }
-        tabBarController?.selectedIndex = 1
+       // tabBarController?.selectedIndex = 1
     }
     
-   
 }
 
 
@@ -139,40 +116,6 @@ extension AddPlantViewController: UIImagePickerControllerDelegate & UINavigation
         dismiss(animated: true, completion: nil)
     }
     
-    
-    
 }
 
-extension Date{
-    var intVal: Int?{
-        if let d = Date.coordinate{
-             let inteval = Date().timeIntervalSince(d)
-             return Int(inteval)
-        }
-        return nil
-    }
-
-    // today's time is close to `2020-04-17 05:06:06`
-
-    static let coordinate: Date? = {
-        let dateFormatCoordinate = DateFormatter()
-        dateFormatCoordinate.dateFormat = "yyyy-MM-dd HH:mm:ss"
-        if let d = dateFormatCoordinate.date(from: "2020-04-17 05:06:06") {
-            return d
-        }
-        return nil
-    }()
-}
-
-
-extension Int{
-    var dateVal: Date?{
-        // convert Int to Double
-        let interval = Double(self)
-        if let d = Date.coordinate{
-            return  Date(timeInterval: interval, since: d)
-        }
-        return nil
-    }
-}
